# 04-Control-Plane-Orchestrator/docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Control Plane Orchestrator API","description":"\n    Job orchestration and management API for Accord Engine.\n    \n    ## Features\n    \n    * **Job Management**: Create, monitor, and cancel jobs\n    * **Queue Management**: Priority-based job queuing with Redis Streams\n    * **Idempotency**: Prevent duplicate job creation\n    * **Status Tracking**: Real-time job status and progress monitoring\n    \n    ## Authentication\n    \n    Currently authentication is disabled for development. In production, use API keys or JWT tokens.\n    ","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/jobs":{"post":{"summary":"Create Job","description":"Create a new job.\n\nArgs:\n    domain: Target domain (e.g., 'amazon.com')\n    url: Target URL\n    job_type: Type of job ('navigate_extract', 'authenticate', etc.)\n    strategy: Execution strategy ('vanilla', 'stealth', 'assault')\n    priority: Priority level (0=emergency, 1=high, 2=normal, 3=low)\n    payload: Job-specific payload data\n    idempotency_key: Optional idempotency key to prevent duplicates\n    timeout_seconds: Job timeout in seconds\n    \nReturns:\n    Job ID and status","operationId":"create_job_api_v1_jobs_post","parameters":[{"name":"domain","in":"query","required":true,"schema":{"type":"string","title":"Domain"}},{"name":"url","in":"query","required":true,"schema":{"type":"string","title":"Url"}},{"name":"job_type","in":"query","required":true,"schema":{"type":"string","title":"Job Type"}},{"name":"strategy","in":"query","required":false,"schema":{"type":"string","default":"vanilla","title":"Strategy"}},{"name":"priority","in":"query","required":false,"schema":{"type":"integer","default":2,"title":"Priority"}},{"name":"idempotency_key","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency Key"}},{"name":"timeout_seconds","in":"query","required":false,"schema":{"type":"integer","default":300,"title":"Timeout Seconds"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"default":{},"title":"Payload"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs/{job_id}":{"get":{"summary":"Get Job Status","description":"Get job status and details.\n\nArgs:\n    job_id: The job ID\n    \nReturns:\n    Job status and details","operationId":"get_job_status_api_v1_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/queue/stats":{"get":{"summary":"Get Queue Stats","description":"Get queue statistics.","operationId":"get_queue_stats_api_v1_queue_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}